requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-readers-web>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0,<1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6
aiofiles>=23.2.0
beautifulsoup4>=4.12.3
lxml>=4.9.0 